refactor: user, auth, post 타입 정리 / community에 있는 const CATEGORIES -> constants/post의 POST_CATEGORIES과 중복되어 삭제 및 기존에 사용부 변경
