+++ .eslintrc.json
@@ -22,6 +22,7 @@
         "argsIgnorePattern": "^_",
         "varsIgnorePattern": "^_"
       }
-    ]
+    ],
+    "@typescript-eslint/no-empty-object-type": "off"
   }
 }
