+++ src/services/authService.ts
@@ -1,13 +1,5 @@
 import { apiClient, authClient } from '@/api/axiosConfig';
-import {
-  EmailVerifyRequest,
-  LoginRequest,
-  OauthRequest,
-  SendCodeRequest,
-  SignupRequest,
-  UserIdCheckRequest,
-  VerifyCodeRequest,
-} from '@/types';
+import { LoginRequest, OauthRequest, SignupRequest } from '@/types';
 
 export const authService = {
   login: async (data: LoginRequest) => {
@@ -17,25 +9,6 @@ export const authService = {
   logout: async () => {
     const response = await authClient.post('/auth/logout');
     authClient.defaults.headers.common['Authorization'] = '';
-    return response.data;
-  },
-  emailVerify: async (data: EmailVerifyRequest) => {
-    const response = await apiClient.post('/auth/email/check', data);
-    return response.data;
-  },
-  sendCode: (data: SendCodeRequest) => {
-    apiClient.post('/auth/email/send', data);
-  },
-  verifyCode: async (data: VerifyCodeRequest) => {
-    const response = await apiClient.post('/auth/email/code', data);
-    return response.data;
-  },
-  userIdCheck: async (data: UserIdCheckRequest) => {
-    const response = await apiClient.post('/auth/userId/check', data);
-    return response.data;
-  },
-  signup: async (data: SignupRequest) => {
-    const response = await apiClient.post('/auth/register', data);
     return response.data;
   },
   oauthCheck: async (data: OauthRequest) => {
