+++ src/services/userService.ts
@@ -1,7 +1,33 @@
-import { authClient } from '@/api/axiosConfig';
-import { ProfileEditFormData } from '@/types';
+import { apiClient, authClient } from '@/api/axiosConfig';
+import {
+  EmailVerifyRequest,
+  ProfileEditFormData,
+  SendCodeRequest,
+  SignupRequest,
+  UserIdCheckRequest,
+  VerifyCodeRequest,
+} from '@/types';
 
 export const userService = {
+  emailVerify: async (data: EmailVerifyRequest) => {
+    const response = await apiClient.post('/auth/email/check', data);
+    return response.data;
+  },
+  sendCode: (data: SendCodeRequest) => {
+    apiClient.post('/auth/email/send', data);
+  },
+  verifyCode: async (data: VerifyCodeRequest) => {
+    const response = await apiClient.post('/auth/email/code', data);
+    return response.data;
+  },
+  userIdCheck: async (data: UserIdCheckRequest) => {
+    const response = await apiClient.post('/auth/userId/check', data);
+    return response.data;
+  },
+  signup: async (data: SignupRequest) => {
+    const response = await apiClient.post('/auth/register', data);
+    return response.data;
+  },
   getUserDrafts: async () => {
     const response = await authClient.get('/user/drafts');
     return response.data;
