+++ src/types/community.ts
@@ -1,18 +1,5 @@
 import { PostStatus } from './post';
 
-export const CATEGORIES = [
-  { id: 0, name: 'Restaurant' },
-  { id: 1, name: 'Shopping' },
-  { id: 2, name: 'Lodging' },
-  { id: 3, name: 'Art' },
-  { id: 4, name: 'Transportation' },
-  { id: 5, name: 'Daily Life' },
-  { id: 6, name: 'Cafe/Dessert' },
-  { id: 7, name: 'Attraction' },
-  { id: 8, name: 'Nature' },
-  { id: 9, name: 'Health' },
-  { id: 10, name: 'Others' },
-] as const;
 export interface Blog {
   id: number;
   title: string;
